Extract common bits, refactor org switching flow
