+++ packages/react/src/components/form/org-switching/useOrgSwitchingFlowState.ts
@@ -1,13 +1,9 @@
 import { useEffect, useMemo, useRef, useState } from "react";
 
 import { useSlashID } from "../../../hooks/use-slash-id";
-import {
-  Flow,
-  createFlow,
-  FlowState,
-  CreateFlowOptions,
-} from "./org-switching-flow";
+import { Flow, createFlow } from "./org-switching-flow";
 import { Handle } from "../../../domain/types";
+import { FlowState, CreateFlowOptions } from "../flow/flow.common";
 
 export function useOrgSwitchingFlowState(opts: CreateFlowOptions) {
   const { recover, sid, __switchOrganizationInContext, user } = useSlashID();
@@ -26,7 +22,7 @@ export function useOrgSwitchingFlowState(opts: CreateFlowOptions) {
       lastUserHandle,
       ...opts,
       logInFn: async () => {
-        return __switchOrganizationInContext({ oid: opts.oid });
+        return __switchOrganizationInContext({ oid: opts.oid! });
       },
       recover,
       cancelFn: () => {
